Change the return value of dw1000_getrangebias() to avoid cast problems Add t_reply_R, t_round_R, t_reply_I, t_round_I to compute the propagation time using Symmetrical approach or Asymmetrical approach fix some "if" condition fix ec_ctrl_val name

+++ dev/dw1000/dw1000-ranging-bias.h
@@ -4,6 +4,6 @@
 
 #include <inttypes.h>
 
-uint16_t dw1000_getrangebias(uint8_t chan, uint16_t range, uint8_t prf);
+uint32_t dw1000_getrangebias(uint8_t chan, uint16_t range, uint8_t prf);
 
 #endif /* _DW1000_RANGING_BIAS_H_ */
